+++ src/main/java/opensearchSDK/ExtensionsRunner.java
@@ -9,7 +9,7 @@
  * GitHub history for details.
  */
 
-package opensearchSDK.transport;
+package opensearchSDK;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
@@ -39,7 +39,8 @@ import org.opensearch.transport.ConnectionManager;
 import org.opensearch.transport.TransportService;
 import org.opensearch.transport.TransportSettings;
 
-import opensearchSDK.netty4.Netty4Transport;
+import opensearchSDK.transport.netty4.Netty4Transport;
+import opensearchSDK.transport.netty4.SharedGroupFactory;
 
 import org.opensearch.transport.TransportInterceptor;
