+++ src/main/java/opensearchSDK/transport/netty4/CopyBytesSocketChannel.java
@@ -44,13 +44,14 @@
  * GitHub history for details.
  */
 
-package opensearchSDK.transport;
+package opensearchSDK.transport.netty4;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelOutboundBuffer;
 import io.netty.channel.RecvByteBufAllocator;
 import io.netty.channel.socket.nio.NioSocketChannel;
+
 import org.opensearch.common.SuppressForbidden;
 import org.opensearch.common.unit.ByteSizeValue;
