+++ src/main/java/opensearchSDK/transport/netty4/Netty4TcpChannel.java
@@ -30,7 +30,7 @@
  * GitHub history for details.
  */
 
-package opensearchSDK.netty4;
+package opensearchSDK.transport.netty4;
 
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelFuture;
